components/telegram/requirements.txt
dalloriam>=0.4.11
google-cloud-error-reporting
requests